Support new option 'use_socket' in sphinx.yml
The option 'use_socket' can be used as a shortcut to enable unix
sockets instead of TCP connections. This is helpfull when running
more than one instance on a shared host as you don't need to set
specific port numbers.

It uses a socket file named 'tmp/sockets/searchd.sock' relative
to the application root.