GroupEnumeratorsMask is part of the defaults.
Otherwise, it only gets added when populated - which is too late for Search's method_missing. Putting populate before mask calls meant scopes couldn't be chained, so this is the best way forward for the moment.

If there were more masks, then perhaps distinguishing them between pre- and post-populate would be worthwhile. Right now, not the case.