Custom database settings can be set in the index definition.
Instead of connecting to a database (though if you're using multiple databases and the schemas are different, that may be required), you can just use set_database in the index definition.

It either takes a hash of database settings with the same keys that Rails expects, or a symbol or string indicating a specific environment (and use the matching configuration from config/database.yml).