Allow for common section in generated configuration for Sphinx 2.2.x.
This is currently only enabled when common_sphinx_configuration: true is set in the appropriate environments in config/thinking_sphinx.yml, as the common section is invalid for Sphinx versions prior to 2.2.x. At some point when Thinking Sphinx expects Sphinx 2.2 as the preferred version, this setting will be enabled by default.

Thanks to @trevorsmith for his contributions to this in both Thinking Sphinx and Riddle.