Use custom join association class where possible.
Essentially, this behaviour only has an impact with Rails 4.1 onwards. We don't need to build a custom scope, but instead can use our join association class instead.

Yes, scopes came in with Rails 4.0, but there's other differences in behaviour that makes removing the scope for 4.0 a pain in the neck.