Allow for different indexing strategies.
TS supplies two: the original behaviour (run the indexer on all indices at once), and the ability to process each index separately.

The latter may eventually become the default, I'm not yet sure.

You can change the setting by putting the following in `config/initializers/thinking_sphinx.rb`:

    ThinkingSphinx::Configuration.instance.indexing_strategy = ThinkingSphinx::IndexingStrategies::OneAtATime
    # or, the default is:
    ThinkingSphinx::Configuration.instance.indexing_strategy = ThinkingSphinx::IndexingStrategies::AllAtOnce

The reason for this is because delta processing requests may happen alongside an all-indices processing, which causes errors in the latter and it stops at that delta index, instead of processing remaining indices. This way, even if a single indexer call fails, the rest will continue.