Prepare all joins at once. Aliases may change.
ActiveRecord's join generation sometimes changes table aliases as you add more joins, and we're using those aliases as we generate the SQL query. Previously, joins were added lazily, but that means we end up with aliases that are no longer correct. Hence now we're prepopulating all association joins to keep things reliable.

This is the underlying cause of #978.