Avoid creating schemas within callbacks.
Rails 5 now fires callbacks when a migration happens, which all well and fine except you end up with nested delta callbacks, both of which are trying to preload the indices. While that call is threadsafe, it can’t be invoked in a recursive manner (the mutex complains).

So, to avoid that, let’s load the extra schema in the spec_helper rather than when the product indices are loaded.