Avoid loading ActiveRecord earlier than necessary.
Related to http://stackoverflow.com/questions/38890554/thinking-spinx-breaks-belongs-to-built-in-validation
Especially in Rails 5 (but maybe in other versions as well), a Rails app will initialise with ActiveRecord::Base present, but ActiveRecord is not completely initialised. This means that some configuration options may not yet be set. Changing this line to wait for the appropriate event means that we can be sure we're on the same page as the app as to when ActiveRecord is actually fired up, and thus we're not skipping any configuration options.