Basic type checking for attribute filters.
Given Riddle now quotes string values in filters, we need to be a little more careful about attribute filter values coming in through params. This is an attempt to do so relatively efficiently.

Parsing all TS index definitions is best avoided until absolutely necessary, hence in this case we’re parsing the configuration file instead.

If you don’t want this middleware in play, you can remove it with the following code in an initialiser:

    ThinkingSphinx::Middlewares::DEFAULT.delete ThinkingSphinx::Middlewares::AttributeTyper
    ThinkingSphinx::Middlewares::RAW_ONLY.delete ThinkingSphinx::Middlewares::AttributeTyper
    ThinkingSphinx::Middlewares::IDS_ONLY.delete ThinkingSphinx::Middlewares::AttributeTyper