Check for deltas at a higher level first.
Only if deltas are enabled and exist for this model do we actually care if there are changes. This does mean things will be slightly slower on the first call (as indices are loaded to the point of knowing if deltas are in play), but that’ll be cached for future calls.