Output a warning when calling ts:index if guard files exist.
To make it easier, we've already got the before_index_hooks ability setup, so it's just a matter of plugging it in to that. Between this and the last commit (9c91644d2c309a16017882820576fb94107dcf71), that covers most situations where indexing occurs and guard files might get in the way. Deltas are the one remaining scenario, and warnings are added to the Rails log for that. Still, documentation needs some work.