Support specs on new MySQL with old Rails versions
When running the older versions of Rails (pre 4.1) with the newer versions of MySQL, MySQL complains because it does not allow null defaults for primary keys, and that’s what Rails tries to do.

This code patches Rails to remove the default NULL value, only for the necessary versions (3.2 and 4.0).