Log exceptions when processing real-time indices, but don't stop.
A good suggestion from @frankmt - though I'd expect this to be a pretty rare situation, and likely not from a bug in Thinking Sphinx itself. If there are exceptions being raised at this point, it's because an instance in the indexed models doesn't know how to handle the data it finds itself with.