Allow calculation of absolute paths.
This is discussed in #816. If you want relative paths translated to absolute paths in the generated configuration, use relative paths in your `config/thinking_sphinx.yml` file, and also add in the setting (per environment) of `absolute_paths: true`.

The default list of configuration options that are paths are set in `ThinkingSphinx::Settings::FILE_KEYS`. If you're using something Sphinx-ish with other settings that should be covered (or Sphinx goes and adds extra settings), you can add to this list by adding the setting `file_keys` (which expects an array of setting names) into your YAML file (again, per environment).