Allow for custom MySQL encodings.
Thinking Sphinx sets Sphinx to automatically use utf8 when talking to MySQL databases. However, MySQL's concept of utf8 is flawed and limited to three bytes, and many characters (included emoji) use more than that. MySQL since 5.5 has utf8mb4 which is 'proper' utf8, and that will become a default in the next significant release of Thinking Sphinx, but for now let's offer the ability to change the default for those who need it. This was raised in #1016.