Use ActiveSupport's lock monitor where possible.
This seems to be related to some changes with Rails' database locking behaviour, which was added in Rails 5.1. This monitor class was added in response to similar issues - and is only present since 5.1.5. The underlying issue noted in #1132 may possibly crop up in earlier Rails 5.1.x releases, but this should resolve things for anything newer. Thanks to @jdelStrother for doing all the hard work on this!