Allow for custom real-time processing.
In particular, this opens the door to parallel processing of indices. Prompted by discussions in #1134.

Use your own Processor class, which can either inherit from `ThinkingSphinx::RealTime::Processor` or be whatever you want, as long as it responds to `call`, accepting an array of indices, and a block that gets called after each index is processed.

https://gist.github.com/pat/53dbe982a2b311f5f7294809109419d2