Delete before insert when transcribing instances.
This was suggested by @njakobsen in #1134, and as he says: “This avoids the need to rebuild the index after adding or removing an index. Without this, the old indexed record would not be deleted since Sphinx's calculation of that record's id would not match now that the number/order of indices has changed.”

It does mean a touch more overhead when transcribing records, but I think it’s worth it.