Direct `reference_name` calls to index_set_class.
This ensures that if IndexSet is customised, that customised class gets this method called as well, rather than the original implementation.

Of course, if `reference_name` isn’t customised, then it’s the original implementation that does get invoked, and that’s fine. This just keeps things consistent - always use whatever configuration’s index_set_class is set to.

This was prompted by thinking through how to support anonymous models as discussed in #1172 by @kalsan.