Specify delta callbacks for associated models.
The same behaviour as what’s offered for real-time indices, but fires off delta callbacks instead, as prompted by #1182.

At this point it’s quite a simple implementation: the `:path` option is used to go back to the associated model via an array of method names, and updates the `delta` column to true for that object. It’s not yet flexible enough to handle delta columns with non-standard names, but it’s enough to prove the concept.