Avoid all app/indices paths from eager loading.
A tweak on the previous commit (02bbb5071482cab8c91b956d2f6437b4cded3cce), this ensures even the app/indices paths in engines are not eagerly loaded by Rails.

Also, it's worth noting that the freezing of eager_load_paths is being done because the original array was frozen. So, trying to leave things as close to how they were found.