fix: ActiveRecord loading, eager_load_paths
There's a few things to note here. Firstly: Rails' loading process
differs in development and production modes, which I suspect is driven
by the eager_load/cache_classes config options.

In development, the `after_initialize` callback is invoked before
ActiveRecord is loaded. In production, it's the other way around. And in
both cases, we need to ensure Thinking Sphinx is included into
ActiveRecord::Base.

So: that insertion now occurs at the end of
lib/thinking_sphinx/active_record.rb - which means that it is only
invoked once, when that file is first required - and we can require it
in both callbacks without needing to care which happens first.

Related to this, is that previously we were only removing `app/indices`
from the eager_load_paths list for Rails 6+. However, that setting
exists in earlier Rails versions as well (I've tested going back to
v4.2) - so, now we're performing this logic across the board. Otherwise:
loading indices when eager loading is enabled (i.e. production) was
occuring before ActiveRecord was loaded (and thus before Thinking Sphinx
is included into ActiveRecord::Base.