fix: update specs to match #1213.
Now that I look a bit deeply, I see there's the distinction
between the number of results known, versus the number that
can be returned (e.g. due to max_matches).

So I'm not sure if we want to keep this change, but let's
see how this evolves. It'd help if I could get Manticore
4.2 working locally!