feat: Allow model/id arguments to Processor.
An evolution of f6657c9d, as per feedback from @akostadinov in #1215. If
you’re dealing with a deleted record, then you may not have an instance
(and cannot retrieve it from the database, as it doesn’t exist). So, the
Processor class now allows for either an instance, or a model and id.

The model should be the actual class, not the string name of the class.