build: GitHub CI.
Now that GitHub Actions allows for retrying of failed jobs only, the
switch is viable. That said, we’re going to keep using Circle CI for
Sphinx 2.2.11, as GitHub no longer supports Ubuntu 16 (and Sphinx v2.2
doesn’t work on anything newer).

The added advantage of this is that we can keep Ruby v2.4, Rails v4.2,
and Bundler v1 to just the Circle CI build, and v2.5+, v5.0+, and v2+
respectively on GitHub.