fix: set a default cutoff for Manticore.
Since Manticore 5.0, their default cutoff is the same as the search
request’s page size - which helps make search requests faster. The
unfortunate trade-off is that it means pagination breaks because we no
longer know what the actual total number of results is (Manticore stops
counting once the cutoff is reached).

So, at least for now, we’re going to set a default cutoff of 0, which
essentially removes that limit.