Allow disabling auto extending ActiveRecord::Base
Method names like `.search` can collide with existing scopes which then
results in the following error:

```
You tried to define a scope named "search" on the model "…", but Active
Record already defined a class method with the same name.
```

With that setting being set to `false` it allows one to enable
ThinkingSphinx per model instead of globally.